requirements.txt
crewai>=0.95.0
anthropic>=0.43.0
pandas>=2.0.0
vaderSentiment>=3.3.2
textblob>=0.18.0
python-dotenv>=1.0.0
